+++ Security/Firewall/Message/LtiPlatformMessageAuthenticationListener.php
@@ -22,21 +22,15 @@ declare(strict_types=1);
 
 namespace OAT\Bundle\Lti1p3Bundle\Security\Firewall\Message;
 
-use Lcobucci\JWT\Parser;
-use OAT\Bundle\Lti1p3Bundle\Security\Authentication\Token\Message\LtiMessageToken;
-use OAT\Library\Lti1p3Core\Message\LtiMessage;
-use OAT\Library\Lti1p3Core\Security\Jwt\AssociativeDecoder;
+use OAT\Bundle\Lti1p3Bundle\Security\Authentication\Token\Message\LtiPlatformMessageSecurityToken;
 use Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface;
-use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Http\Firewall\AbstractListener;
-use Throwable;
 
-class LtiMessageAuthenticationListener extends AbstractListener
+class LtiPlatformMessageAuthenticationListener extends AbstractListener
 {
     /** @var TokenStorageInterface */
     private $storage;
@@ -47,9 +41,6 @@ class LtiMessageAuthenticationListener extends AbstractListener
     /** @var HttpMessageFactoryInterface */
     private $factory;
 
-    /** @var Parser */
-    private $parser;
-
     public function __construct(
         TokenStorageInterface $tokenStorage,
         AuthenticationManagerInterface $authenticationManager,
@@ -58,54 +49,21 @@ class LtiMessageAuthenticationListener extends AbstractListener
         $this->storage = $tokenStorage;
         $this->manager = $authenticationManager;
         $this->factory = $factory;
-        $this->parser = new Parser(new AssociativeDecoder());
     }
 
     public function supports(Request $request): ?bool
     {
-        return null !== $request->get('id_token');
+        return null !== $request->get('JWT');
     }
 
     public function authenticate(RequestEvent $event): void
     {
         $request = $event->getRequest();
 
-        $token = new LtiMessageToken();
+        $token = new LtiPlatformMessageSecurityToken();
         $token->setAttribute('request', $this->factory->createRequest($request));
 
-        try {
-            $this->storage->setToken($this->manager->authenticate($token));
-        } catch (Throwable $exception) {
-            $event->setResponse($this->handleErrorDelegation($exception, $request));
-        }
-    }
+        $this->storage->setToken($this->manager->authenticate($token));
 
-    /**
-     * @throws Throwable
-     */
-    private function handleErrorDelegation(Throwable $exception, Request $request): RedirectResponse
-    {
-        try {
-            $message = new LtiMessage($this->parser->parse($request->get('id_token')));
-        } catch (Throwable $parseException) {
-            throw new AuthenticationException(
-                sprintf('LTI message request authentication failed: %s', $parseException->getMessage()),
-                $parseException->getCode(),
-                $parseException
-            );
-        }
-
-        if (null !== $message->getLaunchPresentation() && null !== $message->getLaunchPresentation()->getReturnUrl()) {
-            $redirectUrl = sprintf(
-                '%s%slti_errormsg=%s',
-                $message->getLaunchPresentation()->getReturnUrl(),
-                strpos($message->getLaunchPresentation()->getReturnUrl(), '?') ? '&' : '?',
-                urlencode($exception->getMessage())
-            );
-
-            return new RedirectResponse($redirectUrl);
-        }
-
-        throw $exception;
     }
 }